Add type inference for query conditions (#719)
This PR modifies the `Query` class so that the query conditions supplied
to its constructor can have their types inferred from the `Schema` the
query is operating against. The following sections will outline the
changes made to support this.

#### `Schema` module and class
- Since queries come from a combination of schema properties and
additional schema dictionaries, a new generic has been added to the
schema class to infer the additional schema dictionaries. This allows
access to those virtual dictionaries in the schema.

- A new `FlattenDocument` helper type was added to the module's exported
types. It will construct an object type that is keyed by the string
keypath (i.e. `foo.bar.baz`) of a document's nested properties with a
value that corresponds to the scalar type at that path.

- A new `InferSchemaPaths` helper type was added to the module's
exported types. It will return a union of the keypaths of its flattened
document. `InferSchemaPaths` isn't technically necessary for this PR,
but it could have utility in other use cases where the keypath is all
that is necessary.

- New exported `DictionaryTypeDefinition...` types were created so that
those could be used for constraint checking.

- The `TSchemaDefinition` generic has been given a default value of
`SchemaDefinition`. This turned out to not cause any ill effects with
inference and allowed significantly easier use of the `Schema` type
elsewhere in the codebase since it wasn't required to supply the generic
value.

#### `Query` module and class
- The `Query` module's `Filter` type has been adjusted to utilize the
`FlattenDocument` type. This type will now produce conditions for each
property in the flattened document which allows for type safety when
constructing query conditions. The `Filter` will now enforce the correct
keys as well as data types for those keys in the conditions, taking into
account the flattened keypaths for the condition keys.

- Additionally, the `Filter` type makes use of a new local
`InferDictionariesType` helper type which will infer the additional
dictionaries supplied to the schema and include those as eligible filter
properties.

- The `SortCriteria` type has been adjusted so that it uses the same
queryable properties as defined by the `Filter` type by extracting its
keys.

#### Other changes
An unrelated change has also been made to eliminate the use of the
`GenericObject` type and replacing it with more appropriate types.

The `Assert` helper type proved to not work properly when attempting to
test the `Filter` utility type. A new much more robust type equality
checking was added to a new `Equals` type module.

Utility types to flatten objects were added to the `Utility` types
module. This was used to produce the flattened document type.