Allow optional properties to be undefined on input for document and model in TS definitions (#732)
When outputting a property defined as optional in a schema, MVOM will
always set that property's value to `null`. The `InferDocumentObject`
and `InferModelObject` utility types correspondingly indicate these
properties as having a type unioned with `null`.

The `InferDocumentObject` type was used in the `DocumentData` type which
is used for the input objects for `Document` and `Model` instances.
However, MVOM does not distinguish between `null` and `undefined` on
input. That is, you do not have to specify an optional property in the
input object. However, the use of the output type `InferDocumentObject`
was forcing specification of optional properties as `null` and did not
allow `undefined`. This PR modifies the `DocumentData` type to leverage
a new utility type named `DeepOptionalNullable`. This utility type will
recurse through an object type and make any properties that are nullable
also optional. This change allows for providing `undefined` for optional
properties for the input data objects when constructing `Model` or
`Document` instances.

Additionally, this PR introduces a new `Remap` utility type which will
make intersection type definitions easier to read by remapping them into
a single object type instead of the intersection type. This was utilized
in the `DeepOptionalNullable` type and was also applied to a few
existing types in the `Schema` module.