Improve the type of schema type path to be a union of tuples (#751)
Currently the type of the `path` property in the `BaseScalarType` is
`number[]`. However, this is misleading as a path is really a numeric
tuple of length 1, 2, or 3.

This PR introduces a new type alias `SchemaTypeDefinitionPath` which was
the type that could be supplied to `SchemaTypeDefinitionBase` as the
`path` property. Additionally, it introduces new type alias
`SchemaTypePath` which is a 1 to 3 length numeric tuple which will be
the type of the `path` property in the `BaseScalarType`. This type is
used in all places where the schema type's path was previously used but
defined as `number[]`.

Finally, the `normalizeMvPath` method in the `BaseScalarType` was
tightened up so that it will ensure that a maximum of 3 positions will
be used in the `path` tuple.