Fix unreported error condition in save subroutine (#753)
In the `save` basic subroutine there is an error check if the `digest`
function fails. However, this code is not properly handling the error
and is merely exiting without finishing the execution of the `save`
subroutine or recording the error.

This PR updates the `save` subroutine so that the error will be properly
recorded.