Add _id and __v to types for schemaless models (#790)
The output of Model methods for models that are constructed without a
schema should have `_id` and `__v` properties as required like models
constructed with a schema. In the current state, the `_id` and `__v`
properties of models without a schema is optional.

This PR augments the `ModelCompositeValue` type so that it marks the
`_id` and `__v` properties as required.