Fix tests that aren't doing what they should (#819)
In the utility type tests within `Schema.test.ts` there is a test for
behavior when `ISOCalendarDateTime` is marked `required`. However, the
test scenario does not mark it as `required` and therefore the test
isn't testing the condition it intended to.

This PR updates the assertion to be working with a `required` schema
property.