Relocate ID checks from Document to Model (#859)
A `Document` instance does not have an `_id` property. This property is
specific to a `Model` instance. However, there exists validations on the
`Document` instance to ensure that the `_id` has the correct pattern
match as well as the construction of the foreign key definitions for the
`_id` property.

This PR relocates the logic for validating an `_id` pattern and building
the `_id` foreign key definitions from the `Document` class to the
`Model` subclass.

Additionally, while doing this work, I noticed that embedded schemas are
not properly building foreign key definitions as no logic was
implemented to do so. The embedded schema type has been updated to build
foreign key definitions.