Update how dates are queried (#907)
### Summary
This PR updates how properties typed as `ISOCalendarDate` are queried on
in an effort to better handle dates that could be in different formats.

#### TypeScript Changes
Updates `ISOCalendarDateTransformer`'s `transformToQuery` method to now
pass in a date query in the `ISOCalendarDate` format (yyyy-MM-dd)
instead of the mv date (days since mv epoch)

#### UniBasic Changes
Updates `find.njk` subroutine to:
1. Capture the system `DATE.FORMAT` at beginning of subroutine 
2. Change `DATE.FORMAT` to `2` (international format) before executing
query
3. Changes `DATE.FORMAT` back to whatever the system is set to after
query executes

### Testing Notes
Tested that a query with a date parameter is now being queried as
`select FOO with BAR >= "2025-04-01"` with date in `yyyy-MM-dd` format.